refactor: move module style files to assets/styles/
